meme_watcher.py
import os
import time
import asyncio
from dotenv import load_dotenv
from solana.rpc.async_api import AsyncClient
from discord_webhook import DiscordWebhook
import json
import logging

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

class MemeWatcher:
    def __init__(self):
        self.solana_client = AsyncClient(os.getenv('SOLANA_RPC_URL'))
        self.discord_webhook_url = os.getenv('DISCORD_WEBHOOK_URL')
        self.min_liquidity = float(os.getenv('MINIMUM_LIQUIDITY_SOL', 100))
        self.price_threshold = float(os.getenv('PRICE_INCREASE_THRESHOLD', 30))
        self.monitoring_interval = int(os.getenv('MONITORING_INTERVAL', 60))
        self.tracked_tokens = {}

    async def send_discord_alert(self, message):
        """Send alert to Discord channel"""
        try:
            webhook = DiscordWebhook(url=self.discord_webhook_url, content=message)
            response = webhook.execute()
            logger.info(f"Discord notification sent: {message}")
            return response
        except Exception as e:
            logger.error(f"Error sending Discord notification: {e}")

    async def check_token_metrics(self, token_address):
        """Analyze token metrics for pump and dump patterns"""
        try:
            # Here you would implement the logic to:
            # 1. Check token liquidity
            # 2. Monitor price movements
            # 3. Check trading volume
            # 4. Look for suspicious patterns
            
            # This is a placeholder for demonstration
            return {
                'is_suspicious': True,
                'reason': 'High price increase with low liquidity'
            }
        except Exception as e:
            logger.error(f"Error checking token metrics: {e}")
            return None

    async def monitor_new_tokens(self):
        """Monitor for new token listings"""
        while True:
            try:
                # Get recent token listings
                # This is where you'd implement the actual token monitoring logic
                # using Solana RPC calls
                
                # Example alert
                await self.send_discord_alert(
                    "🚨 **Potential Pump Detected!**\n"
                    "Token: TEST\n"
                    "Price Increase: 50%\n"
                    "Liquidity: 150 SOL\n"
                    "Risk Level: HIGH"
                )
                
            except Exception as e:
                logger.error(f"Error in monitoring loop: {e}")
            
            await asyncio.sleep(self.monitoring_interval)

    async def start(self):
        """Start the monitoring process"""
        logger.info("Starting Meme Token Watcher...")
        await self.monitor_new_tokens()

if __name__ == "__main__":
    watcher = MemeWatcher()
    asyncio.run(watcher.start())
