requirements.txt
solana==0.30.2
python-dotenv==1.0.0
requests==2.31.0
discord-webhook==1.3.0
web3==6.11.1
python-binance==1.0.19
asyncio==3.4.3
